requirements.txt
aiosmtplib==1.1.2
beautifulsoup4==4.8.2
boto3==1.12.1
botocore==1.15.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
docutils==0.15.2
idna==2.8
jmespath==0.9.4
lxml==4.5.0
pymongo==3.10.1
python-dateutil==2.8.1
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
python-dotenv==0.15.0 